# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bước Vàng Public API
  version: 0.1.0
  description: Public API for the Bước Vàng Zalo Mini App.
  license:
    name: Proprietary
    url: https://buocvang.vn/legal
servers:
  - url: https://api.buocvang.vn
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/challenges:
    get:
      tags: [challenges]
      summary: List challenges.
      operationId: listChallenges
      parameters:
        - name: phase
          in: query
          schema:
            type: string
            enum: [upcoming, live, wrapping]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Challenge list.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Challenge"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [challenges]
      summary: Create a challenge.
      operationId: createChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeRequest"
      responses:
        "201":
          description: Created challenge.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
  /v1/challenges/{id}:
    get:
      tags: [challenges]
      summary: Get a challenge.
      operationId: getChallenge
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: Challenge detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    PaymentRequired:
      description: User does not have enough points.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
        request_id:
          type: string
...
    Challenge:
      type: object
      required: [id, visibility, name, daily_steps_target, duration_days, entry_points, prize_pool, start_date, end_date, status]
      properties:
        id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
          nullable: true
        visibility:
          type: string
          enum: [private, public]
        name:
          type: string
        description:
          type: string
        cover_url:
          type: string
          nullable: true
        daily_steps_target:
          type: integer
        duration_days:
          type: integer
        entry_points:
          type: integer
        prize_pool:
          type: integer
        max_participants:
          type: integer
          nullable: true
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, open, live, settling, settled, cancelled]
        participants:
          type: integer
    CreateChallengeRequest:
      type: object
      required: [visibility, name, daily_steps_target, duration_days, start_date]
      properties:
        visibility:
          type: string
          enum: [private, public]
        name:
          type: string
        description:
          type: string
        cover_url:
          type: string
        daily_steps_target:
          type: integer
          minimum: 1000
          maximum: 50000
        duration_days:
          type: integer
          minimum: 1
          maximum: 90
        start_date:
          type: string
          format: date-time
        entry_points:
          type: integer
          minimum: 0
        max_participants:
          type: integer
          minimum: 1
...